rename palette.js to constants.js
