+++ src/components/AppHeader.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 
-import * as palette from '../palette';
+import { palette } from '../constants';
 
 const StyledHeader = styled.header`
   display: flex;
@@ -22,7 +22,7 @@ const Title = styled.h1`
   color: ${palette.TYPOGRAPHY.heading};
   font-weight: 300;
   font-size: 50px;
-`
+`;
 
 const AppHeader = (props) => (
   <StyledHeader { ...props }>
