+++ src/lib/Menu.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import styled from 'styled-components';
 
-import { palette, style } from '../palette';
+import { palette, style } from '../constants';
 import { NavLink } from './NavLink';
 
 const NavMenu = styled.ul`
