+++ src/lib/StepProgress.js
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 import React, { Component } from 'react';
 import styled from 'styled-components';
 
-import { palette, style } from '../palette';
+import { palette, style } from '../constants';
 
 const StyledStep = styled.div`
   background-color: ${palette.SECONDARY.light};
